Separate sell-related code
